Renaming file and fixing typo
