+++ policy_trainers/burrito_attack.py
@@ -9,8 +9,8 @@ from ray.rllib.algorithms import AlgorithmConfig
 from ray.rllib.algorithms.ppo import PPOConfig
 from ray.tune.logger import pretty_print
 
-from traiderdaive.ray_environments.ray_hyperdrive_env import AGENT_PREFIX, POLICY_PREFIX
 from traiderdaive.ray_environments.attack_hyperdrive_env import AttackHyperdriveEnv
+from traiderdaive.ray_environments.hyperdrive_env import AGENT_PREFIX, POLICY_PREFIX
 from traiderdaive.ray_environments.rewards import TotalRealizedValue
 from traiderdaive.ray_environments.variable_rate_policy import ConstantVariableRate
 from traiderdaive.utils.logging import create_log_dir, get_logger_creator
