+++ policy_trainers/burrito_eval.py
@@ -9,7 +9,7 @@ from ray.rllib.algorithms import AlgorithmConfig
 from ray.rllib.algorithms.ppo import PPOConfig
 
 from traiderdaive.ray_environments.attack_hyperdrive_env import AttackHyperdriveEnv
-from traiderdaive.ray_environments.ray_hyperdrive_env import AGENT_PREFIX, POLICY_PREFIX
+from traiderdaive.ray_environments.hyperdrive_env import AGENT_PREFIX, POLICY_PREFIX
 from traiderdaive.ray_environments.rewards import TotalRealizedValue
 from traiderdaive.ray_environments.variable_rate_policy import ConstantVariableRate
